полный ревью переход на строгую единообразность кода

+++ src/main/java/com/example/fitness_assistant/web/controller/AuthController.java
@@ -4,6 +4,8 @@ import com.example.fitness_assistant.application.service.user.DeleteUserUseCase;
 import com.example.fitness_assistant.application.service.user.LoginResult;
 import com.example.fitness_assistant.application.service.user.LoginUseCase;
 import com.example.fitness_assistant.application.service.user.RegisterUseCase;
+import com.example.fitness_assistant.core.exception.AccessDeniedException;
+import com.example.fitness_assistant.infrastructure.security.UserDetailsAdapter;
 import com.example.fitness_assistant.web.dto.response.AuthResponse;
 import com.example.fitness_assistant.web.dto.request.create.LoginRequest;
 import com.example.fitness_assistant.web.dto.request.create.RegisterRequest;
@@ -11,6 +13,7 @@ import com.example.fitness_assistant.web.mapper.UserWebMapper;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -34,9 +37,14 @@ public class AuthController {
         LoginResult loginResult = loginUseCase.login(request.email(), request.password());
         return userWebMapper.toAuthResponse(loginResult);
     }
+
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Long id) {
+    public void deleteUser(@PathVariable Long id,
+                           @AuthenticationPrincipal UserDetailsAdapter adapter) {
+        if (!adapter.getUserId().equals(id)) {
+            throw new AccessDeniedException();
+        }
         deleteUserUseCase.deleteUser(id);
     }
 }
